Tracker WIP (buggy)
Clients now transfer the segments of owned files so that the tracker knows what the files look like (segments in order) as well as who owns them.

The tracker commit is "buggy" because fileInfo should not have a map of the hashInfos but a vector instead, so that they're kept in order probably.